fix: 🐛 linux not killing

+++ main.js
@@ -165,6 +165,8 @@ const exitPyProc = async () => {
     ]);
   };
 
+  console.log("Killing the process")
+
   await killAllPreviousProcesses();
 
   // check if the platform is Windows
@@ -175,10 +177,16 @@ const exitPyProc = async () => {
     return;
   }
 
+
   // kill signal to pyProc
-  pyflaskProcess.kill();
-  pyflaskProcess = null;
-  PORT = null;
+  if(pyflaskProcess != null) {
+    pyflaskProcess.kill();
+    pyflaskProcess = null;
+  }
+    PORT = null;
+  
+
+
 };
 
 const killAllPreviousProcesses = async () => {
@@ -279,12 +287,13 @@ function initialize() {
               title: "Confirm",
               message: "Any running process will be stopped. Are you sure you want to quit?",
             })
-            .then((responseObject) => {
+            .then(async (responseObject) => {
               let { response } = responseObject;
               if (response === 0) {
                 // Runs the following if 'Yes' is clicked
                 var announcementsLaunch = nodeStorage.getItem("announcements");
                 nodeStorage.setItem("announcements", false);
+                await exitPyProc();
                 quit_app();
               }
             });
@@ -379,6 +388,7 @@ function initialize() {
   });
 
   app.on("window-all-closed", async () => {
+    console.log("All windows closed")
     await exitPyProc();
     app.quit();
   });
